# api/docs/openapi.yaml
openapi: 3.1.0
info:
  title: Colheita API
  version: 1.0.0
  description: API do ciclo PNAE da Cooperativa dos Produtores Rurais de Petrópolis.
servers:
  - url: http://localhost:8080/api/v1
security:
  - bearerAuth: []
paths:
...
  /weeks/{id}/close:
    post: {summary: Fecha uma semana concluída, parameters: [{$ref: '#/components/parameters/Id'}], responses: {'200': {description: Semana fechada}, '409': {$ref: '#/components/responses/Problem'}}}
  /weeks/{id}/reopen:
    post: {summary: Reabre semana com justificativa, parameters: [{$ref: '#/components/parameters/Id'}], responses: {'200': {description: Semana reaberta}}}
  /weeks/{weekId}/{resource}:
    get:
      summary: Lista lançamentos operacionais
      parameters:
        - {name: weekId, in: path, required: true, schema: {type: string, format: uuid}}
        - {name: resource, in: path, required: true, schema: {type: string, enum: [school-orders, school-deliveries, school-returns, producer-allocations, producer-orders, producer-deliveries, producer-returns, weekly-costs]}}
      responses: {'200': {description: Lançamentos}}
  /weeks/{weekId}/summary:
    get: {summary: Balanço financeiro da semana, parameters: [{name: weekId, in: path, required: true, schema: {type: string, format: uuid}}], responses: {'200': {description: Resumo}}}
...
components:
  securitySchemes:
    bearerAuth: {type: http, scheme: bearer, bearerFormat: JWT}
  parameters:
    Id: {name: id, in: path, required: true, schema: {type: string, format: uuid}}
  responses:
    Problem: {description: Erro no formato RFC 9457, content: {application/problem+json: {schema: {type: object}}}}
...